차선 인식 코드 삭제 및 영상 처리 코드 디버깅 및 소켓 통신 코드 완성

+++ restapi/restapi/views.py
@@ -29,11 +29,10 @@ import numpy as np
 import torch
 import shutil
 import logging
+from yt_dlp import YoutubeDL
 from concurrent.futures import ThreadPoolExecutor
 
 from .get_weather_data import get_weather_data
-from .video_utils import download_video_from_url
-from .image_processing_utils import region_of_interest, draw_fit_line, weighted_img, get_fitline, color_filter, dynamic_roi, dynamic_canny, dynamic_hough, grayscale, gaussian_blur, canny, hough_lines
 
 # 로거 설정
 logger = logging.getLogger(__name__)
@@ -46,8 +45,9 @@ logger.addHandler(handler)
 def clean_up_directory(directory, keep_file):
     for filename in os.listdir(directory):
         file_path = os.path.join(directory, filename)
+        print(file_path)
         try:
-            if os.path.isfile(file_path) and file_path != keep_file:
+            if os.path.isfile(file_path) and file_path != keep_file and '_output' not in file_path:
                 os.unlink(file_path)
             elif os.path.isdir(file_path):
                 shutil.rmtree(file_path)
@@ -76,6 +76,12 @@ class ProcessVideoUpload(View):
 
         username = await get_username(request)
 
+        # 모델 로딩을 한 번만 수행하고 재사용합니다.
+        if not hasattr(self, 'model'):
+            device = 'cuda' if torch.cuda.is_available() else 'cpu'
+            self.model = torch.hub.load('ultralytics/yolov5', 'yolov5s', pretrained=True).to(device)
+            self.model.eval()
+
         total_chunks = 0  # 청크 개수를 저장할 변수 초기화
 
         tasks = []
@@ -94,7 +100,7 @@ class ProcessVideoUpload(View):
         completed_tasks = await asyncio.gather(*tasks)
 
         # ThreadPoolExecutor를 사용하여 비디오 프레임 처리를 별도의 스레드에서 실행합니다.
-        executor = ThreadPoolExecutor(max_workers=1)
+        executor = ThreadPoolExecutor(max_workers=2)
         executor.submit(self.run_process_video_frames, f'static/tmp/{username}/{self.random_filename}', self.random_filename, username)
 
         # 응답을 즉시 반환합니다. 프레임 처리는 백그라운드에서 계속 진행됩니다.
@@ -117,15 +123,32 @@ class ProcessVideoUpload(View):
         return file_path  # 파일 경로 반환
 
     async def handle_video_url(self, video_url, username):
-        user_dir = f'tmp/{username}/{video_url.split("/")[-1]}/{video_url.split("/")[-1]}'
+        user_dir = f'static/tmp/{username}'
         os.makedirs(user_dir, exist_ok=True)  # 디렉토리가 이미 존재해도 오류를 발생시키지 않음
-        output_path = f'{user_dir}/{video_url.split("/")[-1]}'
-        await download_video_from_url(video_url, output_path)  # 가정: 비디오 다운로드 함수
-        video_url = await self.process_video_frames(output_path)
-        return video_url  # 이미지 URL 목록 반환
+        random_filename = f"{uuid.uuid4()}.mp4"  # 랜덤한 파일 이름 생성
+        output_path = f'{user_dir}/{random_filename}'
+        
+        # 유튜브 URL에서 비디오 다운로드
+        await self.download_video_from_youtube(video_url, output_path)  # 가정: 유튜브 다운로드 함수
+        
+        return output_path
+
+    async def download_video_from_youtube(self, video_url, output_path):
+        ydl_opts = {
+            'format': 'bestvideo[height<=720]',  # 720p 이하의 최고 화질 비디오만 다운로드
+            'outtmpl': output_path,
+            'postprocessors': [{
+                'key': 'FFmpegVideoConvertor',
+                'preferedformat': 'mp4',  # 비디오 포맷을 mp4로 설정
+            }],
+            'external_downloader_args': {
+                'ffmpeg': ['-t', '180']  # 비디오를 180초(3분)로 제한
+            }
+        }
+        with YoutubeDL(ydl_opts) as ydl:
+            ydl.download([video_url])
 
     def run_process_video_frames(self, path, filename, username):
-        # 새 이벤트 루프를 생성하고 코루틴을 실행합니다.
         new_loop = asyncio.new_event_loop()
         asyncio.set_event_loop(new_loop)
         try:
@@ -137,66 +160,37 @@ class ProcessVideoUpload(View):
         try:
             logger.debug(f"Processing video frames for {filename} in directory {user_dir}")
 
-            # 모델 로딩을 한 번만 수행하고 재사용합니다.
-            if not hasattr(self, 'model'):
-                device = 'cuda' if torch.cuda.is_available() else 'cpu'
-                self.model = torch.hub.load('ultralytics/yolov5', 'yolov5s', pretrained=True).to(device)
-                self.model.eval()
-
-            model = self.model  # 로컬 변수에 할당
-
             reader = imageio.get_reader(f'{user_dir}/{filename}')
             fps = reader.get_meta_data()['fps']
             writer = imageio.get_writer(f'{user_dir}/{filename}_output.mp4', fps=60)
             image_index = 0
+
+            #차선 인식 코드는 불안정하여 제거
+
+            # 클라이언트에 'start' 메시지 전송
+            channel_layer = get_channel_layer()
+            room_group_name = f'process_video_{username}'  # WebSocket 경로 설정
+            await channel_layer.group_send(
+                room_group_name,
+                {
+                    'type': 'video_process_message',
+                    'message': 'start'
+                }
+            )
+
             for i, frame in enumerate(reader):
                 try:
-                    # 프레임 사이즈 조정 로직 추가
-                    height, width = frame.shape[:2]
-                    if height >= 720 and width >= 1280:
-                        frame = cv2.resize(frame, (width * 2 // 3, height * 2 // 3))
-
-                    frame = cv2.cvtColor(frame, cv2.COLOR_RGB2BGR)  # imageio로 읽은 이미지를 cv2 처리 가능한 BGR 포맷으로 변환
-                    height, width = frame.shape[:2]
-                    gray_img = grayscale(frame)
-                    blur_img = gaussian_blur(gray_img, 3)
-                    canny_img = canny(blur_img, 70, 210)
-                    vertices = np.array([[(0, height), (width / 2, height / 2 +100), (width / 2, height / 2 +100), (width, height)]], dtype=np.int32)
-                    ROI_img = region_of_interest(canny_img, vertices)
-                    line_arr = hough_lines(ROI_img, 1, 1 * np.pi / 180, 30, 10, 20)
-                    line_arr = np.squeeze(line_arr)
-                    slope_degree = (np.arctan2(line_arr[:, 1] - line_arr[:, 3], line_arr[:, 0] - line_arr[:, 2]) * 180) / np.pi
-                    line_arr = line_arr[np.abs(slope_degree) < 150]
-                    slope_degree = slope_degree[np.abs(slope_degree) < 150]
-                    line_arr = line_arr[np.abs(slope_degree) > 95]
-                    slope_degree = slope_degree[np.abs(slope_degree) > 95]
-                    L_lines, R_lines = line_arr[(slope_degree > 0), :], line_arr[(slope_degree < 0), :]
-                    if not L_lines.size or not R_lines.size:  # 선 배열이 비어 있는지 확인
-                        print("No lines detected")
-                        return None
-                    temp = np.zeros((frame.shape[0], frame.shape[1], 3), dtype=np.uint8)
-                    L_lines, R_lines = L_lines[:, None], R_lines[:, None]
-                    left_fit_line = get_fitline(frame, L_lines)
-                    right_fit_line = get_fitline(frame, R_lines)
-                    fit_line_xy = np.array([[left_fit_line[0], left_fit_line[1]], [right_fit_line[0],right_fit_line[1]], [left_fit_line[2]+15, left_fit_line[3]+15], [right_fit_line[2]-15,right_fit_line[3]-15]], np.int32)
-                    try:
-                        draw_fit_line(temp, fit_line_xy)
-                    except Exception as e:
-                        logger.error(f"Error drawing fit lines: {e}", exc_info=True)
-                        # 선을 그리지 못했을 경우 로그를 남기고 계속 진행합니다.
-
-                    result = weighted_img(temp, frame)
-                    preds = model(frame)
+                    preds = self.model(frame)
                     preds = preds.pandas().xyxy[0]
                     for index, row in preds.iterrows():
                         if row['confidence'] > 0.4:
                             x1, y1, x2, y2 = int(row['xmin']), int(row['ymin']), int(row['xmax']), int(row['ymax'])
-                            cv2.rectangle(result, (x1, y1), (x2, y2), (0, 255, 0), 2)
-                            cv2.putText(result, f'car {row["confidence"]:.2f}', (x1, y1 - 10), cv2.FONT_HERSHEY_SIMPLEX, 1, (0, 255, 0), 2)
-                    writer.append_data(cv2.cvtColor(result, cv2.COLOR_BGR2RGB))  # 최종 결과를 RGB로 변환하여 writer에 저장
+                            cv2.rectangle(frame, (x1, y1), (x2, y2), (0, 255, 0), 2)
+                            cv2.putText(frame, f'car {row["confidence"]:.2f}', (x1, y1 - 10), cv2.FONT_HERSHEY_SIMPLEX, 1, (0, 255, 0), 2)
+                    writer.append_data(cv2.cvtColor(frame, cv2.COLOR_BGR2RGB))  # 최종 결과를 RGB로 변환하여 writer에 저장
                     if i % 100 == 0:
                         frame_save_path = f'{user_dir}/{filename}_{image_index}.webp'
-                        cv2.imwrite(frame_save_path, cv2.cvtColor(result, cv2.COLOR_BGR2RGB))
+                        cv2.imwrite(frame_save_path, cv2.cvtColor(frame, cv2.COLOR_BGR2RGB))
                         image_index += 1
                 except Exception as e:
                     logger.error(f"Error processing video {filename}: {e}", exc_info=True)
@@ -208,17 +202,15 @@ class ProcessVideoUpload(View):
             
             # 비디오 처리가 완료된 후:
             channel_layer = get_channel_layer()
-            room_group_name = f'ws/process_video/process_video_{username}/'  # WebSocket 경로 설정
-
-            # 메시지 전송
-            async_to_sync(channel_layer.group_send)(
+            room_group_name = f'process_video_{username}'  # WebSocket 경로 설정
+            await channel_layer.group_send(
                 room_group_name,
                 {
                     'type': 'video_process_message',
                     'message': 'completed'
                 }
             )
-            clean_up_directory(user_dir, output_file_path)
+            clean_up_directory(user_dir, f'{user_dir}/{filename}_output.mp4')
             return output_file_path
 
 async def get_username(request):
